pyproject.toml
[project]
name = "excel-agent"
version = "0.1.0"
description = "DeepAgents-based Excel processing agent for data analysis, formula generation, and report creation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywebview>=6.1",
    "langchain-core>=1.2.10,<2.0.0",
    "langchain>=1.2.10,<2.0.0",
    "langchain-anthropic>=1.3.2,<2.0.0",
    "langchain-google-genai>=4.2.0,<5.0.0",
    "wcmatch",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langchain-core>=1.2.15",
    "pywin32>=311; platform_system=='Windows'",
    "fastapi>=0.133.1",
    "uvicorn[standard]>=0.41.0",
    "watchfiles>=1.1.1",
]

[dependency-groups]
dev = [
    "nuitka[onefile]>=4.0.1",
    "pyright>=1.1.408",
    "ruff>=0.15.2",
]

[tool.ruff]
exclude = ["src-python/libs/deepagents"]
